Refactoring, related to timing events

+++ src/lib/gpioEx/sws/Sws.cpp
@@ -132,13 +132,13 @@ void GpioSws::poll() {
   if (active && !readyStage1) { startTimeMs = millis(); readyStage1 = true; VLF("MSG: GpioSws, active"); }
 
   // init delay for synchronization
-  if (readyStage1 && !readyStage2 && (long)(millis() - startTimeMs) > 2000) { readyStage2 = true; VLF("MSG: GpioSws, ready"); }
+  if (readyStage1 && !readyStage2 && millis() - startTimeMs > 2000U) { readyStage2 = true; VLF("MSG: GpioSws, ready"); }
 
   // init timeout
-  if (!readyStage1 && !readyStage2 && (long)(millis() - startTimeMs) > 60000) { lateInitError = true; DLF("ERR: GpioSws, connection failed"); }
+  if (!readyStage1 && !readyStage2 && millis() - startTimeMs > 60000U) { lateInitError = true; DLF("ERR: GpioSws, connection failed"); }
 
   // running timeout
-  if (readyStage2 && (long)(millis() - lastActiveTimeMs) > 5000) { startTimeMs = millis(); lastActiveTimeMs = millis(); active = false; readyStage1 = false; readyStage2 = false; DLF("WRN: GpioSws, connection restart"); }
+  if (readyStage2 && millis() - lastActiveTimeMs > 5000U) { startTimeMs = millis(); lastActiveTimeMs = millis(); active = false; readyStage1 = false; readyStage2 = false; DLF("WRN: GpioSws, connection restart"); }
 }
 
 GpioSws gpio;
