+++ src/lib/tls/ntp/NTP.cpp
@@ -81,14 +81,14 @@ void TlsNTP::poll() {
   Udp.begin(localPort);
 
   // discard any previously received packets
-  unsigned long tOut = millis() + 3000L;
+  unsigned long tOut = millis() + 3000U;
   while ((Udp.parsePacket() > 0) && ((long)(millis() - tOut) < 0)) Y;
 
   VLF("MSG: TLS, transmit NTP Request");
   sendNTPpacket(timeServer);
 
-  uint32_t beginWait = millis();
-  while (millis() - beginWait < 1500) {
+  unsigned long beginWait = millis();
+  while (millis() - beginWait < 1500U) {
     int size = Udp.parsePacket();
     if (size >= NTP_PACKET_SIZE) {
       VLF("MSG: TLS, receive NTP Response");
