+++ src/libApp/temperature/Ds1820.cpp
@@ -93,7 +93,7 @@ void Ds1820::poll() {
 
   if (found) {
     if (index < 0) { requestTime = millis(); if (DS18X20.requestTemperatures(true)) index++; }
-    if ((long)(millis() - requestTime) < 200) return;
+    if (millis() - requestTime < 200U) return;
 
     if (device[index] != (uint64_t)OFF) {
       // loop to read the temperature
@@ -110,7 +110,7 @@ void Ds1820::poll() {
       if (!isnan(temperature)) {
         if (isnan(averageTemperature[index])) averageTemperature[index] = temperature;
         averageTemperature[index] = (averageTemperature[index]*9.0F + temperature)/10.0F;
-        goodUntil[index] = millis() + 30000;
+        goodUntil[index] = millis() + 30000U;
       } else {
         // we must get a reading at least once every 30 seconds otherwise flag the failure with a NAN
         if ((long)(millis() - goodUntil[index]) > 0) averageTemperature[index] = NAN;
